requirements.txt
asgiref==3.2.7
boto3==1.14.18
botocore==1.17.18
Django==3.0.7
django-storages==1.9.1
django-tinymce4-lite==1.8.0
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.10.0
jsmin==3.0.1
Pillow==7.1.2
psycopg2-binary==2.9.5
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
